ghost-system-cloud-full/python/requirements.txt
openai>=1.46.0
markdownify>=0.12.1
python-slugify>=8.0.4
